+++ src/scripts/settings.js
@@ -1,6 +1,7 @@
 var defaultSettings = {
 	'general':{
-		'prereleaseUpdates': false
+		'prereleaseUpdates': false,
+		'stopSounds': true
 	},
 	'playlist':{
 		"soundToBottomAfterPlay": true,
@@ -13,14 +14,22 @@ var defaultSettings = {
 
 // Open settings modal
 function openSettings() {
+	// Set general settings
+	$('#settings-stopSounds').prop('checked', settingsInfo.general.stopSounds);
+	$('#settings-prereleaseUpdates').prop('checked', settingsInfo.general.prereleaseUpdates);
+	// Set playlist settings
 	$('#settings-playlistSoundToBottom').prop('checked', settingsInfo.playlist.soundToBottomAfterPlay);
 	$('#settings-playlistSoundToDelete').prop('checked', settingsInfo.playlist.soundDeleteAfterPlay);
+
+
 	$('#settings-modal').modal('open');
 }
 
 function saveSettings(){
 	settingsInfo.playlist.soundToBottomAfterPlay = $('#settings-playlistSoundToBottom').prop('checked');
 	settingsInfo.playlist.soundDeleteAfterPlay = $('#settings-playlistSoundToDelete').prop('checked');
+	settingsInfo.general.stopSounds = $('#settings-stopSounds').prop('checked');
+	settingsInfo.general.prereleaseUpdates = $('#settings-prereleaseUpdates');
 	storage.storeObj('settings', settings);
 }
 
@@ -37,6 +46,7 @@ function checkSettings(settingsObj) {
 			delete settingsObj[prop];
 		}
 	});
+	console.log(settingsObj);
 }
 
 module.exports = {